[Music Lab] Move blockly code out of MusicView
